working sqlite version w/o MVC

+++ Contacts_Manager.py
@@ -10,27 +10,19 @@ def main():
     try:
         conn = sqlite3.connect('contacts_manager.db')  
         cursor = conn.cursor()
-        print('PRINTING FROM SQLITE3, YAYYY!!!')
 
         while inpt == 'a' or inpt == 'A' or 'r' or inpt == 'R' or inpt == 'l' or inpt == 'L' or 'e' or inpt == 'E' or 'x' or inpt == 'X':
         
             if inpt == 'a' or inpt == 'A':
                 uname, cname, cphone, cemail = add_contact()
 
-                print(uname, cname, cphone, cemail)
 
-                query1 = """SELECT uid FROM users
+                query = """SELECT uid FROM users
                         WHERE uname="{}"
                         """.format(uname)
-                cursor.execute(query1)
+                cursor.execute(query)
                 
                 userid = cursor.fetchone()
-                #userid = int(userid[0])
-                #if type(userid) == int:
-
-
-
-
 
                 if userid == None: #First creates the user and then creates contact when user doesn't exist
                     
@@ -40,13 +32,13 @@ def main():
                             '''.format(uname))
 
                     conn.commit()
-                    print('inserted to users')
 
                     query = """SELECT uid FROM users
                             WHERE uname="{}"
                             """.format(uname)
                     cursor.execute(query)
                     userid = cursor.fetchone()
+                    
 
                     userid = int(userid[0])
 
@@ -58,22 +50,11 @@ def main():
 
                     conn.commit()
 
-                    print('\nContact added.')
+                    print('\n\nContact added.')
 
-
-
-
-
-
-
-
-
-                
-                    
                 
                 elif type(userid[0]) == int: #Directly creates contact when the user already exists
                     userid = int(userid[0])
-                    print('now in if conditin')
                     
                     query = """INSERT INTO contacts(uid,cname,cphone,cemail)
                             VALUES('{}','{}','{}','{}')
@@ -85,50 +66,60 @@ def main():
 
                     print('\nContact added.')
 
-                
-
             #needs changes in WHERE clause
             elif inpt == 'r' or inpt == 'R':
-                uid,cname = edit_contact()
-                query = "DELETE FROM contacts WHERE cname = {}".format(cname)
+                uname,cname = edit_contact()
 
+                query = """SELECT uid FROM users
+                        WHERE uname="{}"
+                        """.format(uname)
                 cursor.execute(query)
-                conn.commit()
+                userid = cursor.fetchone()
 
-                print('\nContact removed.')
+                if userid == None:
+                    print(uname, "doesn't have any contacts.")
+
+                else:
+                    userid = int(userid[0])
+
+                    query = "DELETE FROM contacts WHERE uid = {} and cname = '{}'".format(str(userid),cname)
+
+                    cursor.execute(query)
+                    conn.commit()
+
+                    print('\nContact removed.')
 
             elif inpt == 'l' or inpt == 'L':
                 uname = list_contact()
 
                 query = """SELECT uid FROM users
-                        WHERE uname={}""".format(uname)
+                        WHERE uname="{}"
+                        """.format(uname)
                 cursor.execute(query)
                 userid = cursor.fetchone()
                 
                 if userid != None: 
                     userid = int(userid[0])
-                    #print(userid)
-                    print(uname+"'s contacts:")
+                    print("\n\n"+uname+"'s contacts:")
                 
                     query = """SELECT cname,cphone,cemail
                             FROM contacts
                             WHERE uid={}
-                            ORDER BY cname ASC""".format(userid)
+                            ORDER BY cname ASC""".format(str(userid))
                     cursor.execute(query)
                     rows = cursor.fetchall()
-                    print('hi')
 
-                    print("\n\nList Of All Contacts:")
+                    print("\nList Of All Contacts:")
                     for row in rows:
                         print('\nName:',row[0])
                         
                         phones = row[1].split(' ')
-                        print('Phone(s):')
+                        print('\nPhone(s):')
                         for num in phones:
                             print(num)
                             
                         emails = row[2].split(' ')
-                        print('Email(s):')
+                        print('\nEmail(s):')
                         for email in emails:
                             print(email)
 
@@ -138,7 +129,8 @@ def main():
             elif inpt == 'e' or inpt == 'E':
                 uname,cname = edit_contact()
                 query = """SELECT uid FROM users
-                        WHERE uname={}""".format(uname)
+                        WHERE uname="{}"
+                        """.format(uname)
                 cursor.execute(query)
                 userid = cursor.fetchone()
 
@@ -146,38 +138,32 @@ def main():
                     userid = int(userid[0])
                 
                     query = """SELECT cname, cphone, cemail FROM contacts
-                            WHERE uid = {} and cname={}""".format(userid,cname)
+                            WHERE uid = {} and cname="{}"
+                            """.format(str(userid),cname)
                     cursor.execute(query)
                     row = cursor.fetchone()
 
-                    print('Current Contact:', row[0],row[1],row[2])
+                    print('\nCurrent Contact:')
+                    print('\nName: ', row[0])
+                    print('\nPhone(s): ',row[1])
+                    print('\nEmail(s): ',row[2])
 
                     temp = userid
-                    print(temp)
-                    temp_cname = input('Enter new name: ')
-                    cphone = input('Enter new phone: ')
-                    cemail = input('Enter new email: ')
+                    temp_cname = input('\nEnter new name: ')
+                    cphone = input('\nEnter new phone: ')
+                    cemail = input('\nEnter new email: ')
 
 
-                    print(temp_cname,cphone,cemail,str(temp),cname)
                     
-##                    query = """UPDATE contacts
-##                            SET cname={}, cphone={}, cemail={}
-##                            WHERE uid={} and cname={}""".format(temp_cname,cphone,cemail,str(temp),cname)
-##                
-##                    cursor.execute(query)
-
-                    query = "DELETE FROM contacts WHERE uid={} and cname={}".format(str(temp),cname)
+                    query = "DELETE FROM contacts WHERE uid={} and cname='{}'".format(str(temp),cname)
                     cursor.execute(query)
                     conn.commit()
 
                     query = """INSERT INTO contacts(uid,cname,cphone,cemail)
-                            VALUES({},{},{})""".format(str(temp),temp_cname,cphone,cemail)
+                            VALUES({},'{}','{}','{}')""".format(str(temp),temp_cname,cphone,cemail)
                     cursor.execute(query)
 
                     conn.commit()
-
-
 
                     print('\nContact updated!')
 
@@ -202,31 +188,31 @@ def main():
 
 def header_input():
     print('\n\n\n************CONTACT MANAGER************')
-    print('OPTIONS:')
-    print('Enter (A) to Add a contact')
-    print('Enter (R) to Remove a contact')
-    print('Enter (E) to Edit a contact')
-    print('Enter (L) to List all contacts')
-    print('Enter (X) to Exit application \n')
+    print('\n\nOPTIONS:')
+    print('\nEnter (A) to Add a contact')
+    print('\nEnter (R) to Remove a contact')
+    print('\nEnter (E) to Edit a contact')
+    print('\nEnter (L) to List all contacts')
+    print('\nEnter (X) to Exit application \n')
 
     return input()
 
 
 def add_contact():
-    uname = input("Who's the user?")
-    cname = input("Enter contact name: ")
-    cphone = input("Enter contact phone(s) (If there's more than 1 enter them seperated by a space): ")
-    cemail = input("Enter contact email(s) (If there's more than 1 enter them seperated by a space): ")
+    uname = input("\nEnter user name? ")
+    cname = input("\nEnter contact name: ")
+    cphone = input("\nEnter contact phone (If there's more than 1, enter them seperated by a space): ")
+    cemail = input("\nEnter contact email (If there's more than 1, enter them seperated by a space): ")
     return uname, cname, cphone, cemail
 
 
 def list_contact():
-    uname = input("Enter whose contacts are to be displayed")
+    uname = input("\nEnter whose contacts are to be displayed: ")
     return uname
 
 def edit_contact():
-    uname = input("Enter user name whose contact is to be edited: ")
-    cname = input("Enter name of the contact to be edited: ")
+    uname = input("\nEnter user name whose contact is to be edited/removed: ")
+    cname = input("\nEnter name of the contact to be edited/removed: ")
     return uname,cname
